Animation code refactoring
Some animation code refactoring
