+++ HotScrew/App.js
@@ -1,132 +1,29 @@
 import 'react-native-gesture-handler';
-import React, {useState, useEffect} from 'react';
-import {View, StyleSheet, useWindowDimensions} from 'react-native';
-import Animated, {
-  useSharedValue,
-  useAnimatedStyle,
-  useDerivedValue,
-  useAnimatedGestureHandler,
-  interpolate,
-  withSpring,
-  runOnJS,
-} from 'react-native-reanimated';
-import {PanGestureHandler} from 'react-native-gesture-handler';
+import React from 'react';
+import {View, StyleSheet} from 'react-native';
 
 import Card from './src/components/Profile_Card';
 import users from './assets/data/users';
-import Like from './assets/images/LIKE.png';
-import Nope from './assets/images/nope.png';
 
-const ROTATION = 60;
-const SWIPE_VELOCITY = 800;
+import AnimatedStack from './src/components/Animated_Stack';
 
 const App = () => {
-  const [currentIndex, setCurrentIndex] = useState(0);
-  const [nextIndex, setNextIndex] = useState(currentIndex + 1);
+  const onSwipeLeft = user => {
+    console.warn('swipe left: ', user.name);
+  };
 
-  const currentProfile = users[currentIndex];
-  const nextProfile = users[nextIndex];
-
-  const {width: screenWidth} = useWindowDimensions();
-
-  const hiddenTranslateX = 2 * screenWidth;
-
-  const translateX = useSharedValue(0);
-  const rotate = useDerivedValue(
-    () =>
-      interpolate(translateX.value, [0, hiddenTranslateX], [0, ROTATION]) +
-      'deg',
-  );
-
-  const cardStyle = useAnimatedStyle(() => ({
-    transform: [
-      {
-        translateX: translateX.value,
-      },
-      {
-        rotate: rotate.value,
-      },
-    ],
-  }));
-
-  const nextCardStyle = useAnimatedStyle(() => ({
-    transform: [
-      {
-        scale: interpolate(
-          translateX.value,
-          [-hiddenTranslateX, 0, hiddenTranslateX],
-          [1, 0.8, 1],
-        ),
-      },
-    ],
-    opacity: interpolate(
-      translateX.value,
-      [-hiddenTranslateX, 0, hiddenTranslateX],
-      [1, 0.5, 1],
-    ),
-  }));
-
-  const likeStyle = useAnimatedStyle(() => ({
-    opacity: interpolate(translateX.value, [0, hiddenTranslateX / 5], [0, 1]),
-  }));
-
-  const nopeStyle = useAnimatedStyle(() => ({
-    opacity: interpolate(translateX.value, [0, -hiddenTranslateX / 5], [0, 1]),
-  }));
-
-  const gestureHandler = useAnimatedGestureHandler({
-    onStart: (_, context) => {
-      context.startX = translateX.value;
-    },
-    onActive: (event, context) => {
-      translateX.value = context.startX + event.translationX;
-    },
-    onEnd: event => {
-      if (Math.abs(event.velocityX) < SWIPE_VELOCITY) {
-        translateX.value = withSpring(0);
-        return;
-      }
-
-      translateX.value = withSpring(
-        hiddenTranslateX * Math.sign(event.velocityX),
-        {},
-        () => runOnJS(setCurrentIndex)(currentIndex + 1),
-      );
-    },
-  });
-
-  useEffect(() => {
-    translateX.value = 0;
-    setNextIndex(currentIndex + 1);
-  }, [currentIndex, translateX]);
+  const onSwipeRight = user => {
+    console.warn('swipe right: ', user.name);
+  };
 
   return (
     <View style={styles.pageContainer}>
-      {nextProfile && (
-        <View style={styles.nextCardContainer}>
-          <Animated.View style={[styles.animatedCard, nextCardStyle]}>
-            <Card user={nextProfile} />
-          </Animated.View>
-        </View>
-      )}
-
-      {currentProfile && (
-        <PanGestureHandler onGestureEvent={gestureHandler}>
-          <Animated.View style={[styles.animatedCard, cardStyle]}>
-            <Animated.Image
-              source={Like}
-              style={[styles.like, {left: 10}, likeStyle]}
-              resizeMode="contain"
-            />
-            <Animated.Image
-              source={Nope}
-              style={[styles.like, {right: 10}, nopeStyle]}
-              resizeMode="contain"
-            />
-            <Card user={currentProfile} />
-          </Animated.View>
-        </PanGestureHandler>
-      )}
+      <AnimatedStack
+        data={users}
+        renderItem={({item}) => <Card user={item} />}
+        onSwipeLeft={onSwipeLeft}
+        onSwipeRight={onSwipeRight}
+      />
     </View>
   );
 };
@@ -136,24 +33,6 @@ const styles = StyleSheet.create({
     justifyContent: 'center',
     alignItems: 'center',
     flex: 1,
-  },
-  animatedCard: {
-    width: '90%',
-    height: '70%',
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  nextCardContainer: {
-    ...StyleSheet.absoluteFillObject,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  like: {
-    width: 150,
-    height: 150,
-    position: 'absolute',
-    top: 10,
-    zIndex: 1,
   },
 });
