+++ HotScrew/src/components/Profile_Card/index.js
@@ -30,6 +30,7 @@ const styles = StyleSheet.create({
     width: '100%',
     height: '100%',
     borderRadius: 10,
+    backgroundColor: '#fefefe',
 
     shadowColor: '#000',
     shadowOffset: {
